+++ CHANGELOG.md
@@ -6,12 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Add WebSocket path and port to Session
 ### Changed
 ### Deprecated
 ### Removed
 ### Fixed
 ### Security
+
+## [1.3.1] - 2021-07-20
+### Added
+- Add WebSocket path and port to Session
 
 ## [1.3.0] - 2021-07-20
 ### Added